backend/requirements.txt
requests
python-dotenv
moviepy
gTTS
google-auth-oauthlib
google-auth
google-api-python-client
fastapi
uvicorn